Misc cleanups
Some as suggested by lgtm.com

+++ src/common.cpp
@@ -230,7 +230,7 @@ bool is_windows_subsystem_for_linux() {
     debug_shared(msg_level, L"Backtrace:\n" + join_strings(bt, L'\n') + L'\n');
 }
 
-#else  // HAVE_BACKTRACE_SYMBOLS
+#else   // HAVE_BACKTRACE_SYMBOLS
 
 [[gnu::noinline]] void show_stackframe(const wchar_t msg_level, int, int) {
     debug_shared(msg_level, L"Sorry, but your system does not support backtraces");
@@ -600,12 +600,11 @@ bool should_suppress_stderr_for_tests() {
 static void debug_shared(const wchar_t level, const wcstring &msg) {
     pid_t current_pid;
     if (!is_forked_child()) {
-        std::fwprintf(stderr, L"<%lc> %ls: %ls\n", static_cast<unsigned long>(level), program_name,
-                      msg.c_str());
+        std::fwprintf(stderr, L"<%lc> %ls: %ls\n", level, program_name, msg.c_str());
     } else {
         current_pid = getpid();
-        std::fwprintf(stderr, L"<%lc> %ls: %d: %ls\n", static_cast<unsigned long>(level),
-                      program_name, current_pid, msg.c_str());
+        std::fwprintf(stderr, L"<%lc> %ls: %d: %ls\n", level, program_name, current_pid,
+                      msg.c_str());
     }
 }
 
@@ -1674,7 +1673,7 @@ wcstring format_size(long long sz) {
             if (sz < (1024 * 1024) || !sz_name[i + 1]) {
                 long isz = (static_cast<long>(sz)) / 1024;
                 if (isz > 9)
-                    result.append(format_string(L"%d%ls", isz, sz_name[i]));
+                    result.append(format_string(L"%ld%ls", isz, sz_name[i]));
                 else
                     result.append(
                         format_string(L"%.1f%ls", static_cast<double>(sz) / 1024, sz_name[i]));
